refactor with Claude Code
