+++ .gitignore
@@ -131,5 +131,9 @@ iOSInjectionProject/
 /*.gcno
 **/xcshareddata/WorkspaceSettings.xcsettings
 
+## AI
+
+CLAUDE.md
+
 # End of https://www.toptal.com/developers/gitignore/api/xcode,swift,macos
