+++ .node-version
@@ -1,0 +1,1 @@
+22.14.0
